requirements.txt
pymongo==4.6.0
python-dotenv==0.16.0
openai==1.0.1
pytest==7.4.3